refactor(interp): complete rename profile → prof
- Remove profile/ directory
- interp/interp.go, jit.go, interp_test.go: import updated to
  profpkg "github.com/siyul-park/minivm/prof"
- All profile.Profile / profile.New() references replaced with profpkg equivalents
