+++ interp/interp_test.go
@@ -6,7 +6,7 @@ import (
 	"testing"
 
 	"github.com/siyul-park/minivm/instr"
-	"github.com/siyul-park/minivm/profile"
+	profpkg "github.com/siyul-park/minivm/prof"
 	"github.com/siyul-park/minivm/program"
 	"github.com/siyul-park/minivm/types"
 	"github.com/stretchr/testify/require"
@@ -1860,7 +1860,7 @@ func TestInterpreter_Profile(t *testing.T) {
 	})
 
 	t.Run("injected", func(t *testing.T) {
-		p := profile.New()
+		p := profpkg.New()
 		i := New(program.New(nil), WithProfile(p))
 		defer i.Close()
 		require.Equal(t, p, i.Profile())
